+++ Fpl/FplParser.Tests/TestInFixPostFixPrefix.fs
@@ -1,4 +1,4 @@
-namespace FplParser.Tests
+﻿namespace FplParser.Tests
 
 open FParsec
 open FplParser
@@ -14,7 +14,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfix () =
-        let expected = """"""
         let result = run (definition .>> eof) """def pred Successor postfix "'" (x: Nat) { intr }"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -22,7 +21,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage01 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -30,7 +28,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage01a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x''"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -38,7 +35,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage02 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x(i)'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -46,7 +42,7 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage02a () =
-        let expected = """"""
+        
         let result = run (predicate .>> eof) """x'(i)'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -54,7 +50,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage03 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x[i]'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -62,7 +57,7 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage04 () =
-        let expected = """"""
+        
         let result = run (predicate .>> eof) """x.SomeProperty()'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -70,7 +65,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage05 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """1'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -78,7 +72,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage06a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-f(x)'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -86,7 +79,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage06b () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-f((x + 1))!"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -94,7 +86,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage06c () =
-        let expected = """"""
         let result = run (predicate .>> eof) """f(x)!"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -102,7 +93,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage06d () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-x!"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -110,7 +100,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage07 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """f'(x)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -118,7 +107,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage08 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -126,7 +114,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage08a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x')'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -134,7 +121,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage08b () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x''"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -142,7 +128,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPostfixUsage09 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """x '"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -150,7 +135,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfix () =
-        let expected = """"""
         let result = run (definition .>> eof) """def func Add infix "+" (x,y: Nat) -> Nat { intr }"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -158,7 +142,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfixUsage01 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x + y)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -166,7 +149,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfixUsage01a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x + y + z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -174,7 +156,7 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfixUsage02 () =
-        let expected = """"""
+        
         let result = run (predicate .>> eof) """((x + y) + z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -182,7 +164,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfixUsage03 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x + (y + z))"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -190,15 +171,27 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestInfixUsage04 () =
-        let expected = """"""
-        let result = run (predicate .>> eof) """(x in z)"""
+        let result = run (predicate .>> eof) """(x ∈ z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
         Assert.IsTrue(actual.StartsWith("Success:"))
         
     [<TestMethod>]
+    member this.TestInfixUsage04a () =
+        let result = run (predicate .>> eof) """(x and z)"""
+        let actual = sprintf "%O" result
+        printf "%O" actual
+        Assert.IsTrue(actual.StartsWith("Success:"))
+
+    [<TestMethod>]
+    member this.TestInfixUsage04b () =
+        let result = run (predicate .>> eof) """(x in z)"""
+        let actual = sprintf "%O" result
+        printf "%O" actual
+        Assert.IsTrue(actual.StartsWith("Success:"))
+
+    [<TestMethod>]
     member this.TestCombinedUsage01 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x' + y'' < z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -206,7 +199,7 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage01a () =
-        let expected = """"""
+        
         let result = run (predicate .>> eof) """(-x' + -y'' < -z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -214,7 +207,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage02 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(x' + (y'' < z))"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -222,7 +214,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage03 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """((x' + y'') < z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -230,7 +221,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage04 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(-x' + -y'' + z)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -238,7 +228,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage05 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-(x + -y)'"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -246,7 +235,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage06 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-(x + y)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -254,7 +242,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage07 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-(x + y).Test()"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -262,15 +249,13 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage08 () =
-        let expected = """"""
-        let result = run (predicate .>> eof) """(f -o g)(x)"""
+        let result = run (predicate .>> eof) """(f -∘ g)(x)"""
         let actual = sprintf "%O" result
         printf "%O" actual
         Assert.IsTrue(actual.StartsWith("Success:"))
 
     [<TestMethod>]
     member this.TestCombinedUsage09 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(f + g)'(x)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -278,7 +263,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage10 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(f + -g)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -286,7 +270,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestCombinedUsage10a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """(f + - g)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -294,7 +277,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefix () =
-        let expected = """"""
         let result = run (definition .>> eof) """def func Minus prefix "-" (x: Nat) -> Nat { intr }"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -302,7 +284,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage01 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """'x"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -310,7 +291,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage01a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """''x"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -318,7 +298,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage02 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """'x(i)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -326,7 +305,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage03 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """'x[i]"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -334,7 +312,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage04 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """'x.SomeProperty()"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -342,7 +319,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage05 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """'1"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -350,7 +326,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage06 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-(x)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -358,7 +333,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage06a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-Test(x)"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -366,7 +340,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage07 () =
-        let expected = """"""
         let result = run (predicate .>> eof) """-x"""
         let actual = sprintf "%O" result
         printf "%O" actual
@@ -374,7 +347,6 @@ type TestInfixPostfixPrefix () =
 
     [<TestMethod>]
     member this.TestPrefixUsage07a () =
-        let expected = """"""
         let result = run (predicate .>> eof) """- x"""
         let actual = sprintf "%O" result
         printf "%O" actual